+++ src/kernel/gaussian.hpp
@@ -6,15 +6,15 @@
 class Gaussian: public Kernel {
 public:
     /// @see Kernel::Kernel
-    Gaussian(double h, int N, double mass);
+    Gaussian(float h, int N, float mass);
 
     /// @see Kernel::ValueOf
-    double ValueOf(double r) const;
+    float ValueOf(float r) const;
 
     /// @see Kernel::FOD
-    void FOD(double rx, double ry, double rz, double r, double* ret);
+    void FOD(float rx, float ry, float rz, float r, float* ret);
 
     /// @see Kernel::SOD
-    void SOD(double rx, double ry, double rz, double r, double* ret);
+    void SOD(float rx, float ry, float rz, float r, float* ret);
 };
 #endif // __KERNEL_GAUSSIAN_HPP
